+++ ant-node/src/put_validation.rs
@@ -1259,6 +1259,17 @@ impl Node {
             }
         };
 
+        // Approve the merkle upload when self is lacking of network knowledge.
+        // A normal node shall get K_VALUE (20) entries from the kad get_closet network query.
+        // PEERS_TO_QUERY is currently at the same value as to K_VALUE.
+        if closest_peers.len() < PEERS_TO_QUERY {
+            info!("Only got {} closest peers of {midpoint_address:?}, \
+                   validates the merkle upload request as seems lacking of network knowledge",
+                   closest_peers.len());
+            return Ok(());
+        }
+
+
         let candidate_peer_ids_set: HashSet<_> =
             closest_peers.iter().map(|(peer_id, _)| *peer_id).collect();